+++ src/main/java/io/ezbrains/JpaStarterRead.java
@@ -19,8 +19,17 @@ public class JpaStarterRead {
 
         // or
 
-        AccessCard card = entityManager.find(AccessCard.class,3);
-        System.out.println(card);
+//        AccessCard card = entityManager.find(AccessCard.class,3);
+//        System.out.println(card);
+
+//        PayStub payStub = entityManager.find(PayStub.class,5);
+//        System.out.println(payStub.getEmployee());
+
+        System.out.println("****** Before Fetching employees");
+        Employee employee = entityManager.find(Employee.class,1);
+        System.out.println("****** Before Accessing payStub");
+        System.out.println(employee.getPayStub());
+
 
     }
 }