+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+#  (2023-09-14)
+
+## [Unreleased]
+
+### Added
+- added payStub class
+- added @ManyToOne to payStub and @OneToOne relationship Employee 
+- added PayStub list to employee class
+-
+
 #  (2023-09-09)
 
 ## [Unreleased]